requirements.txt
alembic==1.14.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
email_validator==2.2.0
Flask==3.0.3
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-UUID==0.2
Flask-WTF==1.2.2
gunicorn==23.0.0
humanize==4.10.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==2.1.5
packaging==24.1
pendulum==3.0.0
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
six==1.16.0
SQLAlchemy==2.0.36
text-unidecode==1.3
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
WTForms==3.2.1
